Add extra methods for stack creation and cosulting
- When entering global scope (At the start of the input) no functions
should be called, because method "consulta" already handles creation of
the stack and global symbol table
- When entering a local scope (entering functions, entering command
blocks), the first thing to do would be to push a symbol table to the
stack, where it will contain the information about that scope's
variables.
- When leaving a local scope (Leaving funcion definitions, leaving
command blocks), a pop should be performed on the stack, in order to
remove the symbol table for the closing scope.
- Special case: When declaring functions, it's name and signature should
exist on both the global and local symbol tables, since it could be
called both outside of it's scope (Inside another function) or inside
it's own scope (In case of a recursive function).